ticket #180 import error _make_writeable does not exist in django 1.4-beta
the _make_writeable function has been inlined into TemplateCommand class
in the django 1.4 and up. we should probably rewrite the commands to
use the TemplateCommand class instead of duplicating code. But for now
I moved the _make_writeable function to django_extensions.management.utils
so that the existing commands keep working with Django 1.4 and up.