Fix #329
Removing contribute_to_class doesn't break any existing test and solves
the issue with multi-table inheritance described at #329 and highlighted
by the assertions added in 25b1e260769e418d0e4033e864a726e249dddbe2

contribute_to_class had been added as a workaround for Django
bug https://code.djangoproject.com/ticket/13696 Unfortunately this was
not the right thing to do as it has unwanted side-effects. An auto
field is an implicit primary added automatically by Django.
UUIDField is not an auto field.  The inline editing bug has to be
fixed in Django. django_extensions' UUIDField was doing the right
thing when it wasn't pretending to be an auto field.