sqldiff | Fixed bug with --output_text option and notnull-differ text
Using the --output_text option a bug was raised when a notnull-differ difference
was detected.
The previous output text message for a notnull-differ difference was
"field '%(1)s' null differ: db='%(3)s', model='%(2)s'".
However only 2 parameters were available in diff_args (l.488) ; the name of the model
and whether you should 'SET' or 'DROP' the NULL contraint in the database.