shell_plus: fix handling of extra_args in --notebook
In `run_notebook` after `get_ipython_arguments` is called and the result is
modified, but since both `get_ipython_arguments` and `get_notebook_arguments`
refer to the same list this produces an invalid command line.

Simply returning new lists does not fix the problem because both the
notebook and the kernel fail on invalid flags. This fix handles
extra_args explicitly to ensure the command line is built correctly.